refactor(cmd): extract dispatch and resolution (#525) (#537)
* refactor(cmd): extract dispatch and resolution (#525)

* refactor(cmd): extract completion engine (#526) (#538)
