+++ lua/forge/cmd.lua
@@ -2,8 +2,6 @@ local M = {}
 local collections = require('forge.collections')
 
 local detect = require('forge.detect')
-local ops = require('forge.ops')
-local state_mod = require('forge.state')
 local surface = require('forge.surface')
 
 ---@type table<string, forge.ModifierSpec>
@@ -309,12 +307,6 @@ local function target_parse_opts()
   return require('forge.target').parse_opts()
 end
 
----@param value forge.TargetValue|forge.HeadInput|forge.Scope|nil
----@return forge.RepoLike?
-local function repo_target(value)
-  return require('forge.target').repo_target(value)
-end
-
 local function default_policy(command, _)
   local policy = {}
   if not command.parsed_modifiers.repo and command.family == 'pr' and command.name == 'create' then
@@ -358,428 +350,6 @@ local function default_targets(command, parse_opts)
   end
   return policy, defaults
 end
-
-local function require_git_or_warn()
-  vim.fn.system('git rev-parse --show-toplevel')
-  if vim.v.shell_error ~= 0 then
-    warn('not a git repository')
-    return false
-  end
-  return true
-end
-
-local function require_forge_or_warn()
-  local forge_mod = require('forge')
-  local f = forge_mod.detect()
-  if not f then
-    warn('no forge detected')
-    return nil, forge_mod
-  end
-  return f, forge_mod
-end
-
----@param command forge.Command
----@param forge_name forge.ScopeKind
----@return forge.Scope?
-local function resolve_scope_modifier(command, forge_name)
-  local target = require('forge.target')
-  local repo = repo_target(command.parsed_modifiers.target)
-    or repo_target(command.parsed_modifiers.base)
-    or repo_target(command.parsed_modifiers.head)
-    or repo_target(command.parsed_modifiers.repo)
-    or repo_target(command.default_targets.target)
-    or repo_target(command.default_targets.rev)
-    or repo_target(command.default_targets.base)
-    or repo_target(command.default_targets.head)
-    or repo_target(command.default_targets.repo)
-  return target.resolve_scope(repo, forge_name, target_parse_opts())
-end
-
-local function resolve_repo_modifier(command, forge_name)
-  local target = require('forge.target')
-  local repo = repo_target(command.parsed_modifiers.repo)
-    or repo_target(command.default_targets.repo)
-  return target.resolve_scope(repo, forge_name, target_parse_opts())
-end
-
-local function current_pr_resolution_opts(command, f)
-  local opts = {
-    forge = f,
-  }
-  if command.parsed_modifiers.repo ~= nil then
-    opts.repo = command.parsed_modifiers.repo
-  end
-  if command.parsed_modifiers.head ~= nil then
-    opts.head = command.parsed_modifiers.head
-  end
-  return opts
-end
-
-local function resolve_current_pr_or_warn(command, f)
-  local pr, err = require('forge').current_pr(current_pr_resolution_opts(command, f))
-  if err then
-    warn(err.message)
-    return nil
-  end
-  if pr then
-    return pr
-  end
-  warn(('no open %s found for this branch'):format((f.labels and f.labels.pr_one) or 'PR'))
-  return nil
-end
-
-local function resolve_branch_pr_or_warn(command, f, policy, no_match)
-  local pr, err = require('forge.resolve').branch_pr(current_pr_resolution_opts(command, f), policy)
-  if err then
-    warn(err.message)
-    return nil
-  end
-  if pr then
-    return pr
-  end
-  warn(no_match)
-  return nil
-end
-
-local function dispatch_pr(command)
-  if not require_git_or_warn() then
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local num = command.subjects[1]
-  if command.name == 'create' then
-    local target = require('forge.target')
-    local forge_mod = require('forge')
-    local parse_opts = target_parse_opts()
-    local head = command.parsed_modifiers.head or command.default_targets.head
-    local base = command.parsed_modifiers.base or command.default_targets.base
-    local scope = resolve_repo_modifier(command, f.name)
-    forge_mod.create_pr({
-      draft = command.modifiers.draft == true,
-      instant = command.modifiers.fill == true,
-      web = command.modifiers.web == true,
-      scope = scope,
-      head_branch = head and head.rev or nil,
-      head_scope = target.resolve_scope(head, f.name, parse_opts),
-      base_branch = base and base.rev or nil,
-      base_scope = target.resolve_scope(base, f.name, parse_opts) or scope,
-    })
-    return
-  end
-  if command.name == 'open' then
-    if num then
-      ops.pr_edit({ num = num, scope = resolve_repo_modifier(command, f.name) })
-      return
-    end
-    local pr = resolve_current_pr_or_warn(command, f)
-    if not pr then
-      return
-    end
-    ops.pr_edit(pr)
-    return
-  end
-  if command.name == 'ci' then
-    local pr = num and { num = num, scope = resolve_repo_modifier(command, f.name) }
-      or resolve_branch_pr_or_warn(command, f, {
-        searches = {
-          { 'open' },
-          { 'closed', 'merged' },
-        },
-      }, ('no %s found for this branch'):format((f.labels and f.labels.pr_one) or 'PR'))
-    if not pr then
-      return
-    end
-    ops.pr_ci(f, pr)
-    return
-  end
-  if command.name == 'browse' then
-    local scope = resolve_repo_modifier(command, f.name)
-    if num then
-      ops.pr_browse(f, { num = num, scope = scope })
-    else
-      ops.list_browse(f, 'pr', { scope = scope })
-    end
-    return
-  end
-  if command.name == 'refresh' then
-    state_mod.clear_list_kind('pr')
-    require('forge.logger').info('refreshed ' .. ((f.labels and f.labels.pr) or 'pr') .. ' list')
-    return
-  end
-  local function action_pr()
-    if num then
-      return { num = num, scope = resolve_repo_modifier(command, f.name) }
-    end
-    return resolve_current_pr_or_warn(command, f)
-  end
-  if command.name == 'edit' then
-    ops.pr_edit({ num = num, scope = resolve_repo_modifier(command, f.name) })
-    return
-  end
-  if command.name == 'approve' then
-    local pr = action_pr()
-    if not pr then
-      return
-    end
-    ops.pr_approve(f, pr)
-    return
-  end
-  if command.name == 'merge' then
-    local pr = action_pr()
-    if not pr then
-      return
-    end
-    ops.pr_merge(f, pr, command.modifiers.method)
-    return
-  end
-  if command.name == 'draft' then
-    local pr = action_pr()
-    if not pr then
-      return
-    end
-    ops.pr_toggle_draft(f, pr, false)
-    return
-  end
-  if command.name == 'ready' then
-    local pr = action_pr()
-    if not pr then
-      return
-    end
-    ops.pr_toggle_draft(f, pr, true)
-    return
-  end
-  if command.name == 'close' then
-    local pr = action_pr()
-    if not pr then
-      return
-    end
-    ops.pr_close(f, pr)
-    return
-  end
-  if command.name == 'reopen' then
-    local pr = num and { num = num, scope = resolve_repo_modifier(command, f.name) }
-      or resolve_branch_pr_or_warn(command, f, {
-        searches = {
-          { 'closed' },
-        },
-      }, ('no reopenable %s found for this branch'):format(
-        (f.labels and f.labels.pr_one) or 'PR'
-      ))
-    if not pr then
-      return
-    end
-    ops.pr_reopen(f, pr)
-    return
-  end
-  warn(('unsupported pr action: %s'):format(command.name))
-end
-
-local function dispatch_review(command)
-  if not require_git_or_warn() then
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local num = command.subjects[1]
-  local opts = {
-    adapter = command.modifiers.adapter ~= true and command.modifiers.adapter or nil,
-  }
-  if num then
-    ops.pr_review(f, { num = num, scope = resolve_repo_modifier(command, f.name) }, opts)
-    return
-  end
-  local pr = resolve_current_pr_or_warn(command, f)
-  if not pr then
-    return
-  end
-  ops.pr_review(f, pr, opts)
-end
-
-local function dispatch_issue(command)
-  if not require_git_or_warn() then
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local num = command.subjects[1]
-  local scope = resolve_repo_modifier(command, f.name)
-  if command.name == 'create' then
-    local template = command.modifiers.template
-    require('forge').create_issue({
-      web = command.modifiers.web == true,
-      blank = command.modifiers.blank == true,
-      template = template ~= true and template or nil,
-      scope = scope,
-    })
-    return
-  end
-  if command.name == 'browse' then
-    if num then
-      ops.issue_browse(f, { num = num, scope = scope })
-    else
-      ops.list_browse(f, 'issue', { scope = scope })
-    end
-    return
-  end
-  if command.name == 'edit' then
-    ops.issue_edit({ num = num, scope = scope })
-    return
-  end
-  if command.name == 'close' then
-    ops.issue_close(f, { num = num, scope = scope })
-    return
-  end
-  if command.name == 'reopen' then
-    ops.issue_reopen(f, { num = num, scope = scope })
-    return
-  end
-  if command.name == 'refresh' then
-    state_mod.clear_list_kind('issue')
-    require('forge.logger').info('refreshed issue list')
-    return
-  end
-  warn(('unsupported issue action: %s'):format(command.name))
-end
-
-local function dispatch_ci(command)
-  if not require_git_or_warn() then
-    return
-  end
-  if command.name == 'open' and command.subjects[1] == nil then
-    require('forge').ci({
-      repo = command.parsed_modifiers.repo,
-      head = command.parsed_modifiers.head,
-    })
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local scope = resolve_repo_modifier(command, f.name)
-  if command.name == 'open' then
-    ops.ci_open(f, { id = command.subjects[1], scope = scope })
-    return
-  end
-  if command.name == 'browse' then
-    local id = command.subjects[1]
-    if id then
-      ops.ci_browse(f, { id = id, scope = scope })
-    else
-      ops.list_browse(f, 'ci', { scope = scope })
-    end
-    return
-  end
-  if command.name == 'refresh' then
-    state_mod.clear_list_kind('ci')
-    require('forge.logger').info('refreshed CI run list')
-    return
-  end
-  warn(('unsupported ci action: %s'):format(command.name))
-end
-
-local function dispatch_release(command)
-  if not require_git_or_warn() then
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local tag = command.subjects[1]
-  local scope = resolve_repo_modifier(command, f.name)
-  if command.name == 'browse' then
-    if tag then
-      ops.release_browse(f, { tag = tag, scope = scope })
-    else
-      ops.list_browse(f, 'release', { scope = scope })
-    end
-    return
-  end
-  if command.name == 'delete' then
-    ops.release_delete(f, { tag = tag, scope = scope })
-    return
-  end
-  if command.name == 'refresh' then
-    state_mod.clear_list_kind('release')
-    require('forge.logger').info('refreshed release list')
-    return
-  end
-end
-
-local function dispatch_browse(command)
-  if not require_git_or_warn() then
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local scope = resolve_scope_modifier(command, f.name)
-  local subject = command.subjects[1]
-  local forge_mod = require('forge')
-  if subject then
-    ops.browse_subject(f, { num = subject, scope = scope })
-    return
-  end
-  local location = command.parsed_modifiers.target
-  if location then
-    local branch = location.rev and location.rev.rev or nil
-    if not branch then
-      local explicit_branch = command.parsed_modifiers.branch
-      branch = explicit_branch and explicit_branch.branch or nil
-    end
-    if not branch then
-      local ctx = forge_mod.current_context()
-      branch = type(ctx) == 'table' and ctx.branch or nil
-    end
-    if ops.browse_location(f, location, scope, branch) then
-      return
-    end
-    require('forge.logger').warn('detached HEAD')
-    return
-  end
-  local commit = command.parsed_modifiers.commit
-  if commit and commit.commit then
-    forge_mod.open('browse.commit', { commit = commit.commit, scope = scope })
-    return
-  end
-  local branch = command.parsed_modifiers.branch
-  local file_loc = forge_mod.file_loc(command.range)
-  if branch and branch.branch then
-    if ops.browse_file(f, file_loc, branch.branch, scope) then
-      return
-    end
-    forge_mod.open('browse.branch', { branch = branch.branch, scope = scope })
-    return
-  end
-  local ctx = forge_mod.current_context()
-  local ctx_branch = type(ctx) == 'table' and ctx.branch or nil
-  if ops.browse_file(f, file_loc, ctx_branch, scope) then
-    return
-  end
-  ops.browse_repo({ scope = scope })
-end
-
-local dispatchers = {
-  pr = dispatch_pr,
-  review = dispatch_review,
-  issue = dispatch_issue,
-  ci = dispatch_ci,
-  release = dispatch_release,
-  browse = dispatch_browse,
-  clear = function()
-    state_mod.clear_cache()
-    require('forge.logger').info('cache cleared')
-  end,
-}
 
 ---@param opts? forge.SurfaceNamesOpts
 ---@return string[]
@@ -1057,13 +627,7 @@ function M.parse(args, opts)
 end
 
 function M.dispatch(command)
-  local dispatcher = dispatchers[command.family]
-  if not dispatcher then
-    warn('unknown command: ' .. command.family)
-    return false
-  end
-  dispatcher(command)
-  return true
+  return require('forge.cmd_dispatch').dispatch(command)
 end
 
 function M.run(opts)
@@ -1093,770 +657,8 @@ function M.run(opts)
   return M.dispatch(command)
 end
 
----@param command forge.Command
----@param args string[]
----@return forge.CommandCompletionState
-local function completion_state(command, args)
-  local state = {
-    subjects = {},
-    modifiers = {},
-  }
-  for _, token in ipairs(args) do
-    local name, value
-    local eq = token:find('=', 1, true)
-    if eq then
-      name = token:sub(1, eq - 1)
-      value = token:sub(eq + 1)
-    end
-    if name and collections.list_contains(command.declared_modifiers or {}, name) then
-      if state.modifiers[name] == nil then
-        state.modifiers[name] = value
-      end
-    else
-      state.subjects[#state.subjects + 1] = token
-    end
-  end
-  return state
-end
-
-local function filtered_modifier_completion_items(command, state)
-  local items = {}
-  local names = command.declared_modifiers or command.modifiers
-  for _, name in ipairs(names or {}) do
-    if state.modifiers[name] == nil then
-      local spec = modifiers[name]
-      if spec and spec.kind == 'flag' then
-        items[#items + 1] = name
-      else
-        items[#items + 1] = name .. '='
-      end
-    end
-  end
-  return items
-end
-
-local function filter(candidates, arglead)
-  return vim.tbl_filter(function(s)
-    return s:find(arglead, 1, true) == 1
-  end, candidates)
-end
-
-local function system_lines(cmd)
-  if type(cmd) == 'table' then
-    local result = vim.system(cmd, { text = true }):wait()
-    if result.code ~= 0 then
-      return {}
-    end
-    local output = vim.trim(result.stdout or '')
-    return output == '' and {} or vim.split(output, '\n', { plain = true, trimempty = true })
-  end
-  local lines = vim.fn.systemlist(cmd)
-  if vim.v.shell_error ~= 0 then
-    return {}
-  end
-  return lines
-end
-
-local function add_completion_candidate(items, seen, value)
-  if type(value) ~= 'string' or value == '' or seen[value] then
-    return
-  end
-  seen[value] = true
-  items[#items + 1] = value
-end
-
-local function repo_completion_values(prefix)
-  local target = require('forge.target')
-  local parse_opts = target_parse_opts()
-  local items = {}
-  local seen = {}
-
-  local alias_names = {}
-  for name in pairs(parse_opts.aliases or {}) do
-    alias_names[#alias_names + 1] = name
-  end
-  table.sort(alias_names)
-  for _, name in ipairs(alias_names) do
-    add_completion_candidate(items, seen, name)
-  end
-
-  for _, remote in ipairs(system_lines({ 'git', 'remote' })) do
-    add_completion_candidate(items, seen, remote)
-    local resolved = target.resolve_repo(remote, parse_opts)
-    if resolved then
-      add_completion_candidate(items, seen, resolved.slug)
-      if resolved.host and resolved.slug then
-        add_completion_candidate(items, seen, resolved.host .. '/' .. resolved.slug)
-      end
-    end
-  end
-
-  for _, repo in ipairs({
-    target.current_repo(parse_opts),
-    target.push_repo(parse_opts),
-    target.collaboration_repo(parse_opts),
-  }) do
-    if repo then
-      add_completion_candidate(items, seen, repo.slug)
-      if repo.host and repo.slug then
-        add_completion_candidate(items, seen, repo.host .. '/' .. repo.slug)
-      end
-    end
-  end
-
-  return filter(items, prefix)
-end
-
-local function ref_completion_values(prefix)
-  local items = {}
-  local seen = {}
-  for _, ref in
-    ipairs(system_lines('git for-each-ref --format=%(refname:short) refs/heads refs/tags'))
-  do
-    add_completion_candidate(items, seen, ref)
-  end
-  for _, sha in
-    ipairs(system_lines({ 'git', 'rev-list', '--max-count=20', '--abbrev-commit', 'HEAD' }))
-  do
-    add_completion_candidate(items, seen, sha)
-  end
-  return filter(items, prefix)
-end
-
-local function rev_address_completion_values(prefix)
-  local items = {}
-  local seen = {}
-  local at = prefix:find('@', 1, true)
-  if at then
-    local repo = prefix:sub(1, at - 1)
-    local rev_prefix = prefix:sub(at + 1)
-    local base = repo ~= '' and (repo .. '@') or '@'
-    for _, ref in ipairs(ref_completion_values(rev_prefix)) do
-      add_completion_candidate(items, seen, base .. ref)
-    end
-    return items
-  end
-
-  for _, repo in ipairs(repo_completion_values(prefix)) do
-    add_completion_candidate(items, seen, repo .. '@')
-  end
-  for _, ref in ipairs(ref_completion_values('')) do
-    add_completion_candidate(items, seen, '@' .. ref)
-  end
-  return filter(items, prefix)
-end
-
-local function target_completion_values(prefix)
-  if prefix:find(':', 1, true) then
-    return {}
-  end
-  local items = {}
-  local seen = {}
-  local at = prefix:find('@', 1, true)
-  if at then
-    for _, value in ipairs(rev_address_completion_values(prefix)) do
-      add_completion_candidate(items, seen, value .. ':')
-    end
-    return items
-  end
-
-  for _, repo in ipairs(repo_completion_values(prefix)) do
-    add_completion_candidate(items, seen, repo .. '@')
-  end
-  for _, ref in ipairs(ref_completion_values('')) do
-    add_completion_candidate(items, seen, '@' .. ref .. ':')
-  end
-  return filter(items, prefix)
-end
-
-local completion_limits = {
-  pr = 100,
-  issue = 100,
-  ci = 30,
-  release = 30,
-}
-
-local function json_list(cmd)
-  if type(cmd) ~= 'table' then
-    return nil
-  end
-  local result = vim.system(cmd, { text = true }):wait()
-  if result.code ~= 0 then
-    return nil
-  end
-  local ok, data = pcall(vim.json.decode, result.stdout or '[]')
-  if not ok or type(data) ~= 'table' then
-    return nil
-  end
-  return data
-end
-
-local function scoped_id(id, suffix)
-  if suffix ~= nil and suffix ~= '' then
-    return id .. '|' .. suffix
-  end
-  return id
-end
-
-local function completion_scope_key(forge_mod, scope)
-  if type(forge_mod.scope_key) == 'function' then
-    return forge_mod.scope_key(scope)
-  end
-  return ''
-end
-
-local function completion_limit(forge_mod, kind)
-  local cfg = type(forge_mod.config) == 'function' and forge_mod.config() or nil
-  local display = type(cfg) == 'table' and cfg.display or nil
-  local limits = type(display) == 'table' and display.limits or nil
-  if kind == 'pr' then
-    return type(limits) == 'table' and limits.pulls or completion_limits.pr
-  end
-  if kind == 'issue' then
-    return type(limits) == 'table' and limits.issues or completion_limits.issue
-  end
-  if kind == 'ci' then
-    return type(limits) == 'table' and limits.runs or completion_limits.ci
-  end
-  if kind == 'release' then
-    return type(limits) == 'table' and limits.releases or completion_limits.release
-  end
-  return 50
-end
-
-local function completion_scope(forge_mod, f, state)
-  local repo = state.modifiers.repo
-  if type(repo) == 'string' and repo ~= '' then
-    local target = require('forge.target')
-    local resolved = target.resolve_repo(repo, target_parse_opts())
-    if resolved then
-      return target.repo_scope(resolved, f.name)
-    end
-  end
-  if type(forge_mod.current_scope) == 'function' then
-    return forge_mod.current_scope(f.name)
-  end
-  return nil
-end
-
-local function completion_forge(state)
-  local ok, forge_mod = pcall(require, 'forge')
-  if not ok or type(forge_mod) ~= 'table' or type(forge_mod.detect) ~= 'function' then
-    return nil, nil, nil
-  end
-  local f = forge_mod.detect()
-  if not f then
-    return nil, forge_mod, nil
-  end
-  return f, forge_mod, completion_scope(forge_mod, f, state)
-end
-
-local function completion_list_key(forge_mod, kind, state, scope)
-  return state_mod.list_key(kind, scoped_id(state, completion_scope_key(forge_mod, scope)))
-end
-
-local function cached_completion_list(forge_mod, kind, states, scope)
-  local items = {}
-  local found = false
-  for _, state in ipairs(states) do
-    local cached = state_mod.get_list(completion_list_key(forge_mod, kind, state, scope))
-    if type(cached) == 'table' then
-      found = true
-      for _, item in ipairs(cached) do
-        items[#items + 1] = item
-      end
-    end
-  end
-  if found then
-    return items
-  end
-  return nil
-end
-
-local function fetch_completion_list(forge_mod, f, kind, state, scope)
-  local limit = completion_limit(forge_mod, kind)
-  local cmd
-  if kind == 'pr' and type(f.list_pr_json_cmd) == 'function' then
-    cmd = f:list_pr_json_cmd(state, limit, scope)
-  elseif kind == 'issue' and type(f.list_issue_json_cmd) == 'function' then
-    cmd = f:list_issue_json_cmd(state, limit, scope)
-  elseif kind == 'ci' and type(f.list_runs_json_cmd) == 'function' then
-    cmd = f:list_runs_json_cmd(state == 'all' and nil or state, scope, limit)
-  elseif kind == 'release' and type(f.list_releases_json_cmd) == 'function' then
-    cmd = f:list_releases_json_cmd(scope, limit)
-  end
-  local data = json_list(cmd)
-  if type(data) ~= 'table' then
-    return {}
-  end
-  if kind == 'ci' and type(f.normalize_run) == 'function' then
-    local normalized = {}
-    for _, item in ipairs(data) do
-      normalized[#normalized + 1] = f:normalize_run(item)
-    end
-    data = normalized
-  end
-  state_mod.set_list(completion_list_key(forge_mod, kind, state, scope), data)
-  return data
-end
-
-local function completion_list(forge_mod, f, kind, states, fetch_state, scope)
-  return cached_completion_list(forge_mod, kind, states, scope)
-    or fetch_completion_list(forge_mod, f, kind, fetch_state or states[1], scope)
-end
-
-local function completion_entry(value)
-  return { value = value }
-end
-
----@param value string?
----@param forge_name forge.ScopeKind
----@return forge.Scope?
-local function completion_repo_scope(value, forge_name)
-  if type(value) ~= 'string' or value == '' then
-    return nil
-  end
-  return require('forge.target').resolve_scope(value, forge_name, target_parse_opts())
-end
-
----@param value forge.RepoLike?
----@return forge.Scope?
-local function completion_repo_like_scope(value)
-  if type(value) ~= 'table' or value.kind == 'repo' then
-    return nil
-  end
-  ---@cast value forge.Scope
-  return value
-end
-
----@param value string?
----@return forge.RevTarget?
-local function completion_head(value)
-  if type(value) ~= 'string' or value == '' then
-    return nil
-  end
-  return require('forge.target').parse_rev(value, target_parse_opts())
-end
-
----@param state forge.CommandCompletionState
----@param f forge.Forge
----@param include_head boolean?
----@return forge.CurrentPROpts?
-local function completion_current_pr_opts(state, f, include_head)
-  local opts = {
-    forge = f,
-  }
-  if state.modifiers.repo ~= nil then
-    local scope = completion_repo_scope(state.modifiers.repo, f.name)
-    if not scope then
-      return nil
-    end
-    opts.repo = scope
-  end
-  if include_head ~= false and state.modifiers.head ~= nil then
-    local head = completion_head(state.modifiers.head)
-    if not head then
-      return nil
-    end
-    opts.head = head
-  end
-  return opts
-end
-
----@param pr forge.PRRef
----@param state string
----@param pr_state forge.PRState?
----@return forge.PickerEntry
-local function pr_completion_entry(pr, state, pr_state)
-  ---@type forge.PRCompletionValue
-  local value = {
-    num = pr.num,
-    scope = pr.scope,
-    state = state,
-  }
-  if type(pr_state) == 'table' then
-    value.is_draft = pr_state.is_draft
-    value.review_decision = pr_state.review_decision
-    value.mergeable = pr_state.mergeable
-  end
-  return completion_entry(value)
-end
-
----@param state forge.CommandCompletionState
----@return forge.PRCompletionTarget?
-local function implicit_pr_completion_target(state)
-  local forge_mod = require('forge')
-  local f = forge_mod.detect()
-  if not f then
-    return nil
-  end
-  local opts = completion_current_pr_opts(state, f)
-  if not opts then
-    return nil
-  end
-  ---@type forge.PRRef?
-  local pr
-  ---@type forge.CmdError?
-  local err
-  pr, err = forge_mod.current_pr(opts)
-  if err then
-    return nil
-  end
-  if pr then
-    local pr_state = state_mod.pr_state(f, pr.num, pr.scope)
-    local state_name = type(pr_state) == 'table' and pr_state.state or nil
-    return {
-      forge = f,
-      entry = pr_completion_entry(pr, state_name or 'OPEN', pr_state),
-    }
-  end
-  pr, err = require('forge.resolve').branch_pr(opts, {
-    searches = { { 'closed' } },
-  })
-  if err then
-    return nil
-  end
-  if pr then
-    return {
-      forge = f,
-      entry = pr_completion_entry(pr, 'CLOSED'),
-    }
-  end
-  pr, err = require('forge.resolve').branch_pr(opts, {
-    searches = { { 'merged' } },
-  })
-  if err then
-    return nil
-  end
-  if pr then
-    return {
-      forge = f,
-      entry = pr_completion_entry(pr, 'MERGED'),
-    }
-  end
-  return nil
-end
-
----@param command forge.Command
----@param state forge.CommandCompletionState
----@return forge.PRCompletionTarget?
-local function merge_method_completion_target(command, state)
-  if command.family ~= 'pr' or command.name ~= 'merge' then
-    return nil
-  end
-  local forge_mod = require('forge')
-  local f = forge_mod.detect()
-  if not f then
-    return nil
-  end
-  local opts = completion_current_pr_opts(state, f, false)
-  if not opts then
-    return nil
-  end
-  local num = state.subjects[1]
-  if num then
-    local scope = completion_repo_like_scope(opts.repo)
-    local pr_state = state_mod.pr_state(f, num, scope)
-    local state_name = type(pr_state) == 'table' and pr_state.state or nil
-    return {
-      forge = f,
-      entry = pr_completion_entry({ num = num, scope = scope }, state_name or 'OPEN', pr_state),
-    }
-  end
-  local pr, err = forge_mod.current_pr(opts)
-  if err or not pr then
-    return nil
-  end
-  local pr_state = state_mod.pr_state(f, pr.num, pr.scope)
-  local state_name = type(pr_state) == 'table' and pr_state.state or nil
-  return {
-    forge = f,
-    entry = pr_completion_entry(pr, state_name or 'OPEN', pr_state),
-  }
-end
-
----@param command forge.Command?
----@param state forge.CommandCompletionState
----@param values string[]
----@return string[]
-local function filter_family_verb_completion_items(command, state, values)
-  if not command then
-    return values
-  end
-  local items = {}
-  local target = nil
-  if command.family == 'pr' and command.name == 'open' then
-    target = implicit_pr_completion_target(state)
-  end
-  local policy = require('forge.completion_policy')
-  for _, value in ipairs(values) do
-    local forge = target and target.forge or nil
-    local entry = target and target.entry or nil
-    if policy.verb(command, value, forge, entry) then
-      items[#items + 1] = value
-    end
-  end
-  return items
-end
-
-local function complete_pr_subjects(command, state, prefix, policy)
-  local f, forge_mod, scope = completion_forge(state)
-  if not f or not forge_mod then
-    return {}
-  end
-  local prs =
-    completion_list(forge_mod, f, 'pr', policy.states_to_consult, policy.fetch_state, scope)
-  local fields = f.pr_fields or {}
-  local items = {}
-  local seen = {}
-  for _, pr in ipairs(prs or {}) do
-    local num = tostring(pr[fields.number] or '')
-    local entry = completion_entry({
-      num = num,
-      scope = scope,
-      state = pr[fields.state],
-      is_draft = fields.is_draft and pr[fields.is_draft] or nil,
-    })
-    if not policy.available or policy.available(command.name, f, entry) then
-      add_completion_candidate(items, seen, num)
-    end
-  end
-  return filter(items, prefix)
-end
-
-local function complete_issue_subjects(command, state, prefix, policy)
-  local f, forge_mod, scope = completion_forge(state)
-  if not f or not forge_mod then
-    return {}
-  end
-  local issues =
-    completion_list(forge_mod, f, 'issue', policy.states_to_consult, policy.fetch_state, scope)
-  local fields = f.issue_fields or {}
-  local items = {}
-  local seen = {}
-  for _, issue in ipairs(issues or {}) do
-    local num = tostring(issue[fields.number] or '')
-    local entry = completion_entry({
-      num = num,
-      scope = scope,
-      state = issue[fields.state],
-    })
-    if not policy.available or policy.available(command.name, f, entry) then
-      add_completion_candidate(items, seen, num)
-    end
-  end
-  return filter(items, prefix)
-end
-
-local function complete_run_subjects(state, prefix, policy)
-  local f, forge_mod, scope = completion_forge(state)
-  if not f or not forge_mod then
-    return {}
-  end
-  local runs =
-    completion_list(forge_mod, f, 'ci', policy.states_to_consult, policy.fetch_state, scope)
-  local items = {}
-  local seen = {}
-  for _, run in ipairs(runs or {}) do
-    add_completion_candidate(items, seen, tostring(run.id or ''))
-  end
-  return filter(items, prefix)
-end
-
-local function complete_release_subjects(state, prefix, policy)
-  local f, forge_mod, scope = completion_forge(state)
-  if not f or not forge_mod then
-    return {}
-  end
-  local releases =
-    completion_list(forge_mod, f, 'release', policy.states_to_consult, policy.fetch_state, scope)
-  local fields = f.release_fields or {}
-  local items = {}
-  local seen = {}
-  for _, release in ipairs(releases or {}) do
-    add_completion_candidate(items, seen, tostring(release[fields.tag] or ''))
-  end
-  return filter(items, prefix)
-end
-
----@param command forge.Command
----@param state forge.CommandCompletionState
-local function completion_values(command, state, flag_name, prefix)
-  local spec = modifiers[flag_name]
-  local policy = require('forge.completion_policy').modifier_value(command, flag_name, spec)
-  if not policy then
-    return nil
-  end
-  if policy.source == 'repo' then
-    return repo_completion_values(prefix or '')
-  end
-  if policy.source == 'ref' then
-    return ref_completion_values(prefix or '')
-  end
-  if policy.source == 'rev_address' then
-    return rev_address_completion_values(prefix or '')
-  end
-  if policy.source == 'target' then
-    return target_completion_values(prefix or '')
-  end
-  if policy.source == 'template' then
-    return filter(require('forge').template_slugs(), prefix or '')
-  end
-  if policy.source == 'adapter' then
-    return filter(require('forge').review_adapter_names(), prefix or '')
-  end
-  if policy.source == 'available_merge_methods' then
-    local target = merge_method_completion_target(command, state)
-    if not target then
-      return filter(spec.values, prefix or '')
-    end
-    return filter(
-      require('forge.availability').pr_merge_methods(target.forge, target.entry),
-      prefix or ''
-    )
-  end
-  if policy.source == 'command_values' then
-    return filter(command.modifier_values[flag_name], prefix or '')
-  end
-  if policy.source == 'modifier_values' then
-    return filter(spec.values, prefix or '')
-  end
-  return nil
-end
-
-local function subject_completion_items(command, state, arglead)
-  local subject = command.subject or { min = 0, max = 0 }
-  local max = subject.max or subject.min or 0
-  if max ~= nil and #state.subjects >= max then
-    return {}
-  end
-  if subject.kind == 'branch' then
-    return ref_completion_values(arglead)
-  end
-  if subject.kind == 'rev' then
-    return ref_completion_values(arglead)
-  end
-  if subject.kind == 'sha' then
-    return filter(
-      system_lines({ 'git', 'rev-list', '--max-count=20', '--abbrev-commit', 'HEAD' }),
-      arglead
-    )
-  end
-  local policy = require('forge.completion_policy').subject(command)
-  if policy.cmdline_usefulness == 'suppress' then
-    return {}
-  end
-  if not policy.allow_empty_prefix and arglead == '' then
-    return {}
-  end
-  if policy.subject_kind == 'pr_number' then
-    return complete_pr_subjects(command, state, arglead, policy)
-  end
-  if policy.subject_kind == 'issue_number' then
-    return complete_issue_subjects(command, state, arglead, policy)
-  end
-  if policy.subject_kind == 'ci_run_id' then
-    return complete_run_subjects(state, arglead, policy)
-  end
-  if policy.subject_kind == 'release_tag' then
-    return complete_release_subjects(state, arglead, policy)
-  end
-  return {}
-end
-
 function M.complete(arglead, cmdline, _)
-  local words = split_words(cmdline)
-  local arg_idx = arglead == '' and #words or #words - 1
-  local family_name = words[2]
-  local surface_opts = {
-    forge_name = detect.forge_name(),
-  }
-  local family = M.family(family_name, surface_opts)
-  local explicit_verb = family
-      and words[3] ~= nil
-      and (family.verbs[words[3]] ~= nil or (family.aliases and family.aliases[words[3]] ~= nil))
-      and words[3]
-    or nil
-
-  local flag, value_prefix = arglead:match('^([%w%-_]+)=(.*)$')
-  if flag then
-    local command = M.resolve(family_name, explicit_verb, surface_opts)
-    if command then
-      command.declared_modifiers = command.modifiers or {}
-      local rest_index = explicit_verb and 4 or 3
-      local consumed = {}
-      for i = rest_index, #words - 1 do
-        consumed[#consumed + 1] = words[i]
-      end
-      local state = completion_state(command, consumed)
-      local values = completion_values(command, state, flag, value_prefix)
-      if values then
-        return vim.tbl_map(function(v)
-          return flag .. '=' .. v
-        end, values)
-      end
-    end
-  end
-  if arg_idx == 1 then
-    return filter(
-      M.family_names({
-        include_aliases = true,
-        forge_name = surface_opts.forge_name,
-      }),
-      arglead
-    )
-  end
-  if not family then
-    return {}
-  end
-  if arg_idx == 2 then
-    local command = M.resolve(family_name, nil, surface_opts)
-    local state = command and completion_state(command, {}) or { modifiers = {}, subjects = {} }
-    local slot_policy = require('forge.completion_policy').family_slot(command)
-    local candidates = {}
-    if slot_policy.include_verbs then
-      for _, verb in
-        ipairs(
-          filter_family_verb_completion_items(
-            command,
-            state,
-            M.verb_names(family_name, surface_opts)
-          )
-        )
-      do
-        candidates[#candidates + 1] = verb
-      end
-    end
-    if command then
-      if slot_policy.include_modifiers then
-        vim.list_extend(candidates, filtered_modifier_completion_items(command, state))
-      end
-      if slot_policy.include_subjects then
-        vim.list_extend(candidates, subject_completion_items(command, state, arglead))
-      end
-    end
-    return filter(candidates, arglead)
-  end
-  local command = M.resolve(family_name, explicit_verb, surface_opts)
-  if not command then
-    return {}
-  end
-  command.declared_modifiers = command.modifiers or {}
-  local rest_index = explicit_verb and 4 or 3
-  local consumed = {}
-  for i = rest_index, arglead == '' and #words or (#words - 1) do
-    consumed[#consumed + 1] = words[i]
-  end
-  local state = completion_state(command, consumed)
-  local slot_policy = require('forge.completion_policy').argument_slot(command, state)
-  local candidates = {}
-  if slot_policy.include_modifiers then
-    vim.list_extend(candidates, filtered_modifier_completion_items(command, state))
-  end
-  if slot_policy.include_subjects then
-    vim.list_extend(candidates, subject_completion_items(command, state, arglead))
-  end
-  return filter(candidates, arglead)
+  return require('forge.cmd_complete').complete(M, arglead, cmdline, split_words)
 end
 
 return M
